# Speedup MVT column parser