--- a/postgis/mvt.h
+++ b/postgis/mvt.h
@@ -46,6 +46,15 @@
 
 #include "vector_tile.pb-c.h"
 
+typedef struct mvt_column_cache
+{
+    uint32_t *column_keys_index;
+    uint32_t *column_oid;
+    Datum *values;
+    bool *nulls;
+    TupleDesc tupdesc;
+} mvt_column_cache;
+
 typedef struct mvt_agg_context
 {
 	char *name;
@@ -67,6 +76,7 @@ typedef struct mvt_agg_context
 	uint32_t values_hash_i;
 	uint32_t keys_hash_i;
 	uint32_t row_columns;
+	mvt_column_cache column_cache;
 } mvt_agg_context;
 
 /* Prototypes */