@Algunenano 2018-08-10
Travis is [having some issues](https://www.traviscistatus.com/incidents/pxj3py3xzn14) with its network and some builds are failing, but since 3/4 went ok I'll proceed anyway.